+++ alarm.py
@@ -1,8 +1,11 @@
 #!/usr/bin/python3
 
 import sys
+import datetime
+
 from PyQt6.QtWidgets import QMainWindow, QPushButton, QApplication, QLabel, QLineEdit, QDialog
 from PyQt6.QtCore import QTimer, Qt
+from PyQt6.QtGui import QFont
 
 class TimeDialog(QDialog):
     def __init__(self, parent):
@@ -47,9 +50,14 @@ class TimeDialog(QDialog):
 
     def _sendInfo(self):
         # Retrieve number of hours, minutes, and seconds inputted and init timer
+
+        # TODO: Check input so hours can only be 0-99, minutes: 0-59, seconds: 0-59
         self.parent.hours = self.hoursIn.text()
         self.parent.minutes = self.minutesIn.text()
         self.parent.seconds = self.secondsIn.text()
+
+        self.parent._initTimerValues()
+
         self.close()
 
 
@@ -74,9 +82,12 @@ class Window(QMainWindow):
         self.setTimeButton.clicked.connect(self._getTime)
 
         # Create label to show time left
-        self.label = QLabel("0", self)
-        self.label.setGeometry(100, 200, 200, 50)
+        self.label = QLabel("00:00:00", self)
+        self.label.setGeometry(50, 200, 300, 100)
         self.label.setAlignment(Qt.AlignmentFlag.AlignCenter)
+
+        # Formal time label
+        self.label.setFont(QFont('Helvetica', 64))
 
         # Create start button
         self.startButton = QPushButton("Start", self)
@@ -99,9 +110,9 @@ class Window(QMainWindow):
         self.resetButton.clicked.connect(self._resetTimer)
 
         # Create test button
-        self.resetButton = QPushButton("Test", self)
-        self.resetButton.setGeometry(125, 550, 150, 40)
-        self.resetButton.clicked.connect(self._testValues)
+        # self.resetButton = QPushButton("Test", self)
+        # self.resetButton.setGeometry(125, 550, 150, 40)
+        # self.resetButton.clicked.connect(self._testValues)
 
         # Create timer
         self.timer = QTimer(self)
@@ -109,42 +120,64 @@ class Window(QMainWindow):
         self.timer.start(100)
 
     def _getTime(self):
+        # Triggered when "Set time" button pressed
         self.start = False
 
         # Init other window
         self.timeDialog.show()
 
+    def _initTimerValues(self):
+        # Get total seconds from "set time" dialog
+        totalTime = int(self.hours) * 3600 + int(self.minutes) * 60 + int(self.seconds)
+        self.count = totalTime * 10
+        self.lastVal = self.count / 10
+
+        self._setTimer(self.lastVal)
+
     def _startTimer(self):
+        # Start timer
         self.start = True
 
         if self.count == 0:
             self.start = False
     
     def _pauseTimer(self):
+        # Pause timer
         self.start = False
 
     def _restartTimer(self):
+        # Restart timer using last used hours, minutes, seconds values
         self.count, self.start = self.lastVal * 10, False
-        self.label.setText(str(self.lastVal))
+        self._setTimer(self.lastVal)
 
     def _resetTimer(self):
+        # Reset timer and all values
         self.count, self.start, self.lastVal = 0, False, 0
-        self.label.setText("0")
+        self.hours, self.minutes, self.seconds = 0, 0, 0
+        self.label.setText("00:00:00")
 
     def _testValues(self):
+        # Debug function
         print(self.hours, self.minutes, self.seconds)
 
     def _showTimer(self):
+        # Used to update timer
         if self.start:
             self.count -= 1
 
             if self.count == 0:
                 self.start = False
-                self.label.setText("0")
+                self.label.setText("00:00:00")
                 # TODO: Sound alarm!
         
         if self.start:
-            self.label.setText(str(self.count / 10))
+            self._setTimer(self.count / 10)
+    
+    def _setTimer(self, secs_in):
+        # Used to format timer
+        dt_curr = datetime.timedelta(seconds=round(secs_in))
+        dt_curr = int(dt_curr.total_seconds())
+        self.label.setText('{:02}:{:02}:{:02}'.format(dt_curr // 3600, dt_curr % 3600 // 60, dt_curr % 60))
 
 if __name__ == '__main__':
     app = QApplication([])