+++ HISTORY.md
@@ -2,7 +2,18 @@
 This document outlines all notable changes to this project thus far. Inspired by https://keepachangelog.com/en/1.0.0/.
 
 ## [Unreleased]
-- Timer is currently being worked on. Latest version (0.0.2) is not functional.
+- Input parsing
+- Alarm when timer is done
+
+## [0.1.0] - 2022-10-29
+### Added
+- Created basic timer in form HH:MM:SS
+
+### Changed
+- UI, specifically of the timer
+
+### Fixed
+- Connected time entry dialog properly with main window
 
 ## [0.0.2] - 2022-10-28
 ### Added
@@ -10,8 +21,8 @@ This document outlines all notable changes to this project thus far. Inspired by
 - Added a restart button to restart the timer with the current values
 
 ### Changed
-- Updated UI
-- Changed name of "Reset" button to "Stop"
+- UI
+- Name of "Reset" button to "Stop"
 
 ## [0.0.1] - 2022-10-27
 ### Added